fix: separate azstorage and pgtileserv metadata json
