+++ sites/geohub/src/routes/api/vector/azstorage/metadata.json/+server.ts
@@ -1,0 +1,4 @@
+import api from '$api';
+import type { RequestEvent, RequestHandler } from './$types';
+
+export const GET: RequestHandler = (event: RequestEvent) => api.handle(event);
